refactor(pageTemplate): create detailedDataverseItem component
